Update README.md
The primary objective of this project is to design and develop a Power BI dashboard that not only captivates visually but also efficiently presents key insights.

The aim is to provide stakeholders with a tool that facilitates informed decision-making by offering strategic insights in a clear and understandable manner. Through this dashboard, stakeholders will be empowered to decipher complex data effortlessly, thus enabling them to make well-informed strategic decisions with ease.